Random tile generation works with images now
Instead of just creating a dungeon of random tiles of a certain colour,
now a dungeon of random tiles corresponding to a certain image is now
generated.

In addition, the map constants (width, tile_size etc, image surfaces)
are placed in their own module to be used as a sort of "header" file.
Also, the code has been grouped into classes to make it easier to extend
later on when procedural dungeon generation will be done.