Generalized room generation code, made into a class method
The room generation code was generalized since it was only specific for
generating a room within the confines of the dungeon. When eventually
implementing BSP to generate the dungeon, a more generic function is
required since the dungeon will be split into several sub-dungeons.
Therefore, a class method of the Room class was created. The method can
generate a room enclosed within any arbitrary area given certain
attributes of said area.