Room placement when generating dungeons works but has flaws.
First of all, the borders of the rooms collide. This is likely due to
the fact that the rooms should placed within a region that is big enough
to contain the room and a region of void around the room. This will
ensure that no two adjacent rooms in the dungeon will collide and that a
corrider can be drawn between them.

Also, I don't like how sometimes a generated dungeon is very sparsely
populated with rooms. Not entirely sure how to deal with it though.