Fixed room placement within dungeon generation algorithm
With this new version of the algorithm, the placed rooms are spaced out
such that they don't collide and have enough space to fit a hallway
between them. In addition, the old code had some off by one errors so
the visualizer lines would overlap each other which was bad. Now all the
dividing lines do not overlap and the generated rooms fall within the
boundaries provided by these lines, which proves that the algorithm is
working as intended. Note that some of these improvements were obtained
by being more precise in the definition of the Room and Dungeon objects
such that the invariant properties of an instance of each object make it
so that room collision can never happen.

Now what's next is to create hallways between each of the room such that
the entire dungeon is connected.