Connecting individual rooms works but dungeon connection doesn't
Connecting two isolated rooms works however connecting the entire
dungeon fails. It seems that when connecting rooms in the dungeon
generation algorithm, two rooms that can be connected together
horizontally or vertically will blindly connect, overwriting rooms
inbetween them. This looks horrible. In addition, if a room cannot be
connected horizontally or vertically, then the algorithm doesn't connect
them. Perhaps, one way to circumvent this would be to find the two rooms
that are the closest together and then try to connect the two closest
rooms together. This could potentially be expensive (O(n^2) for the
trivial solution, probably O(nlogn) for a more optimized solution).
Another alternative would be to use a pathfinding algorithm such as A*
search or Djikstra's to connect the two rooms. This could lead to some
interesting looking hallways.