Tried to connect rooms using greedy first search, extremely hacky
Well I hoped that it would connect all the rooms but it often finds no
paths and the paths it does find are sometimes not even possible. I
tried using a greedy best-first search instead of Djikstra's or A*
search since it was faster. Now just have to either make this algorithm
work properly or use different one.