Dungeon connection works but is very ugly
So connecting the dungeon works as in every room is connected to at
least one other room and thus the player can reach everyplace in the
dungeon. However, the generated hallways will draw over existing rooms
and stuff. The way to get around this would be to use a pathfinder like
A* search, Djikstra's, or even Jump Point Search. JPS is a bit to
complicated to implement and we only have a week left so a more standard
pathfinder like A* or Djikstra's will have to be used. However, once the pathfinder has been implemented, we can use reuse the logic in the connect_rooms() method so that it returns the start and end points of the hallway and run the pathfinder on these points.

In the unlikely event that
adding the pathfinder still doesn't resolve this issue, then we can
create a BSP tree explicitly and have the Dungeon.generate_dungeon()
method return this BSP tree. Then, we can change connect_rooms()
within the Dungeon class so that will recursively traverse this tree after the dungeon has been generated using a
DFS and then connect every child to its sister.