Room connection mostly works now, but still not guaranteed to not collide
Changed room connection logic to find the two neighbouring rooms that
are the minimum distance from each other and then connect them. For the
most part this seems to make sure that the hallways do not overwrite the
rooms. However, in the case that two neighbouring rooms that are very
close together and have to be connected using an l-shaped path,
collisions can still occur.