Generated borders around hallways now works, changed tilemap dict
I added a method in the dungeon class that loops over all the hallways
in the dungeon and adds borders around them. Also, when I made this
method I noticed that I made a mistake when creating the tilemap, I
swapped the rows and the cols. So I fixed it. Also, I CHANGED THE
TILEMAP TO A DICT. This was really important, because I needed to
update all the tiles in a hallway's path with the dungeon's tilemap.
Converting the tilemap to a dict was the cleanest way to do this. So
when accessing the tilemap instead of doing tilemap[row][col] you have
to use tilemap[(col, row)]. So instead of a 2D array, the tilemap is a
dictionary of positions (tuples). I changed the Djikstra's algorithm you
built slightly because I think your earlier version had to account for
my mistake with the tilemap. I tested it and it works fine. But all the
old 2d array code was left commented out just in case.