Added random grass tiles, cleaned up level.py, and game over sound
The code for level.py was pretty poorly commented and had a lot of
functions that were not necessary anymore, so I cleaned that up. Added
drawing random grass tiles within the dungeon to make it look less
homogenous. Could expand this concept to mud and puddles and other sorts
of stuff. Lastly, I added a game over sound.